+++ src/scrapper/main/resources/client.conf
@@ -3,5 +3,5 @@
   term: "camiones"
   lat: 21.85717833
   long: -102.28487238
-  quantity: 250
+  quantity: 500
 }
